requirements.txt
Django==5.1.3
mysql-connector-python==9.1.0
Pillow==10.4.0
qrcode[pil]==7.4.2
mysqlclient==2.2.0